refactor(frontend:userprofile): make profile screen compatible with the rest of the navigation.
