+++ frontend/types/user.ts
@@ -1,0 +1,14 @@
+export type User = {
+    country: string,
+    display_name: string,
+    email: string,
+    explicit_content: { filter_enabled: boolean, filter_locked: boolean },
+    external_urls: { spotify: string },
+    followers: { href: string, total: number },
+    href: string,
+    id: number,
+    images: [{ height: number, url: string, width: number }],
+    product: string,
+    type: string,
+    uri: string
+}
